requirements.txt
flask==2.0.2
flask-cors==3.0.10
psycopg2-binary==2.9.1
